Began work on MSVC Build Scripts
THIS IS NOT DONE YET, nor is it close, but at the end, SkyvoCommon will be buildable with MSVC.

+++ Common/math/include/Constants.h
@@ -5,7 +5,7 @@
 
 namespace Math{
 
-const static float PI = M_PI;
+const static float PI = static_cast<float>(M_PI);
 const static float HALF_PI = PI / 2.0f;
 const static float QUARTER_PI = PI / 4.0f;
 const static float EIGHTH_PI = PI / 8.0f;
